+++ src/main/java/frc/robot/RobotContainer.java
@@ -16,8 +16,8 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import frc.robot.commands.DeployPneumatics;
-import frc.robot.commands.DriveFieldOriented;
-import frc.robot.commands.DriveRobotOriented;
+import frc.robot.commands.Drive.DriveFieldOriented;
+import frc.robot.commands.Drive.DriveRobotOriented;
 import frc.robot.commands.EnterXMode;
 import frc.robot.commands.RunBelts;
 import frc.robot.commands.RunIntake;
